+++ src/errorHandlers/handleHttpClientErrors.ts
@@ -1,21 +1,36 @@
-import {HttpErrorResponse} from '@angular/common/http';
+import {Notifications} from '@anglr/common';
 
 import {HttpClientError} from '../misc/classes/httpClientError';
-import {readErrorsFromHttpErrorResponse} from '../misc/utils';
-import {HttpClientErrorOptions} from '../interfaces';
+import {CatchHttpClientErrorOptions} from '../interfaces';
+import {CLIENT_ERROR_NOTIFICATIONS} from '../misc/tokens';
+import {ServerValidationService} from '../services';
+import {CatchHttpClientErrorBehavior} from '../misc/enums';
+import {applyBehavior} from '../misc/utils';
 
 /**
- * Handles http error response with code 400..499 and returns HttpClientError, otherwise returns original http error response
+ * Handles http client errors according provided options
+ * @param error - Instance of `HttpClientError`
+ * @param options - Options used for handling errors
  */
-export async function handleHttpClientErrors(error: HttpErrorResponse, options: HttpClientErrorOptions): Promise<HttpClientError|HttpErrorResponse>
+export async function handleHttpClientErrors(error: HttpClientError, options: CatchHttpClientErrorOptions): Promise<HttpClientError|null>
 {
-    //handles client codes (400..499)
-    if(error.status >= 400 && error.status < 500)
-    {
-        const {errors, validationErrors} = await readErrorsFromHttpErrorResponse(error, options.injector, options.clientErrorsResponseMapper, options.clientValidationErrorsResponseMapper);
+    const notifications: Notifications|null = options.injector.get(CLIENT_ERROR_NOTIFICATIONS, null, {optional: true});
+    const behavior: CatchHttpClientErrorBehavior = options.behavior ?? CatchHttpClientErrorBehavior.Suppress;
 
-        return new HttpClientError(errors, validationErrors, error);
+    //show server validations errors
+    if(error.validationErrors)
+    {
+        options.injector?.get(ServerValidationService).addServerValidationErrors(error.validationErrors);
     }
 
-    return error;
+    //show client errors
+    if(!options.skipErrorNotifications && notifications)
+    {
+        for(const clientError of error.errors)
+        {
+            notifications.error(clientError);
+        }
+    }
+
+    return applyBehavior(error, behavior);
 }