+++ src/errorHandlers/index.ts
@@ -1,4 +1,5 @@
 export * from './handle400WithValidationsFunc';
 export * from './handle404Func';
 export * from './handle4xxFunc';
+export * from './handleHttpClientErrorResponse';
 export * from './handleHttpClientErrors';