+++ src/index.ts
@@ -10,6 +10,7 @@ export * from './misc/classes/httpNotFoundError';
 export * from './misc/classes/internalServerErrorInfo';
 export * from './misc/classes/restClientError';
 export * from './misc/classes/restNotFoundError';
+export * from './misc/enums';
 export * from './misc/providers';
 export * from './misc/tokens';
 export * from './misc/types';