+++ src/interfaces/index.ts
@@ -1,5 +1,6 @@
 export * from './anglrExceptionExtender/anglrExceptionExtender.interface';
 export * from './angularError/angularError.interface';
+export * from './catchHttpClientErrorOptions/catchHttpClientErrorOptions.interface';
 export * from './errorWithStack/errorWithStack.interface';
 export * from './errorWithUrl/errorWithUrl.interface';
 export * from './handle4xxOptions/handle4xxOptions.interface';