+++ src/rxjsOperators/index.ts
@@ -1,4 +1,5 @@
 // export * from './handle400WithValidations.operator';
 // export * from './handle404.operator';
 // export * from './handle4xx.operator';
-export * from './processHttpClientError.operator';
+export * from './catchHttpClientError.operator';
+export * from './processHttpClientErrorResponse.operator';